Made vairables pod local, not monitor local
