--- RasterPropMonitor/Core/IPageElement.cs
@@ -1,7 +1,0 @@
-namespace JSI
-{
-    internal interface IPageElement
-    {
-        void HandlePageCreate(RasterPropMonitor propMonitor);
-    }
-}
